pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kaggle-math"
version = "0.1.0"
description = "AIMO3 Nemotron Nano vLLM-first solver system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
runtime = [
  "datasets>=3.0.0",
  "openai>=1.40.0",
  "torch>=2.2.0",
  "transformers>=4.57.0",
  "vllm>=0.12.0",
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
testpaths = ["tests"]
